docs: rename according to camelcase
