+++ src/components/pages/startPage.ts
